+++ examples/cmake_project/Findzarrs.cmake
@@ -20,11 +20,11 @@ FetchContent_Declare(zarrs_ffi_source
 FetchContent_Populate(zarrs_ffi_source)
 
 # Fetch the dependencies at configure time using cargo fetch
-# get_property(CARGO_EXECUTABLE TARGET Rust::Cargo PROPERTY IMPORTED_LOCATION)
-# execute_process(COMMAND ${CARGO_EXECUTABLE} fetch WORKING_DIRECTORY ${zarrs_ffi_source_SOURCE_DIR})
+get_property(CARGO_EXECUTABLE TARGET Rust::Cargo PROPERTY IMPORTED_LOCATION)
+execute_process(COMMAND ${CARGO_EXECUTABLE} fetch WORKING_DIRECTORY ${zarrs_ffi_source_SOURCE_DIR})
 
 # Build zarrs_ffi, creates a zarrs_ffi target aliased to zarrs::zarrs
-corrosion_import_crate(MANIFEST_PATH ${zarrs_ffi_source_SOURCE_DIR}/Cargo.toml FEATURES ${zarrs_FIND_COMPONENTS}) #  FLAGS --offline
+corrosion_import_crate(MANIFEST_PATH ${zarrs_ffi_source_SOURCE_DIR}/Cargo.toml FEATURES ${zarrs_FIND_COMPONENTS} FLAGS --frozen)
 # corrosion_experimental_cbindgen(TARGET zarrs_ffi HEADER_NAME "zarrs.h") # not working
 
 # add zarrs.h to include directories