+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - Add `Cargo.lock` to version control
 
+### Changed
+ - Use `--frozen` in `Findzarrs.cmake`
+
 ## [0.8.2] - 2024-08-13
 
 ### Added